Move selectCivs to Civs
